requirements.txt
# =================================================================
# FRAUD DETECTION ENGINE - DEPENDENCY CONFLICT RESOLVED
# =================================================================

# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Data Validation
pydantic==2.6.1
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Caching
redis==5.0.1
aioredis==2.0.1

# Data Processing - CONFLICT RESOLVED
pandas==2.1.4
numpy==1.26.4
scipy==1.12.0

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0
imbalanced-learn==0.12.0
joblib==1.3.2

# MLOps - VERSION DOWNGRADED TO AVOID CONFLICT
mlflow==2.8.1
optuna==3.5.0

# Message Queue
kafka-python==2.0.2
celery[redis]==5.3.6
flower==2.0.1

# Monitoring & Observability
prometheus-client==0.20.0
structlog==24.1.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.2

# HTTP Client
httpx==0.27.0
aiohttp==3.9.3

# Utilities
python-dotenv==1.0.1
click==8.1.7
rich==13.7.0
typer==0.9.0

# Data Visualization
plotly==5.18.0
matplotlib==3.8.3
seaborn==0.13.2

# Configuration
pyyaml==6.0.1
toml==0.10.2

# Async Support
asyncio-mqtt==0.16.2
aiofiles==23.2.1

# Time & Date - CONFLICT RESOLVED
python-dateutil==2.8.2
pytz==2023.3

# Validation & Serialization
marshmallow==3.21.0
cerberus==1.3.5

# Performance
orjson==3.9.15
ujson==5.9.0

# Email
aiosmtplib==3.0.1
jinja2==3.1.3

# File Processing
openpyxl==3.1.2
xlsxwriter==3.2.0

# Networking
dnspython==2.6.1

# Build Dependencies
setuptools==69.1.0
wheel==0.42.0
